# Performance improvement in creating the extra index