--- a/hdt-java-core/src/main/java/org/rdfhdt/hdt/hdt/impl/HDTImpl.java
+++ b/hdt-java-core/src/main/java/org/rdfhdt/hdt/hdt/impl/HDTImpl.java
@@ -509,9 +509,9 @@ public class HDTImpl implements HDTPrivate {
 
 			// SAVE
 			if(this.hdtFileName!=null) {
-				FileOutputStream out=null;
+				BufferedOutputStream out=null;
 				try {
-					out = new FileOutputStream(versionName);
+					out = new BufferedOutputStream(new FileOutputStream(versionName));
 					ci.clear();
 					triples.saveIndex(out, ci, listener);
 					out.close();