@D063520 2022-01-12
I worked on this too. I'm for these changes. To summarize:
1) we add a buffered stream to write to the file
2) It looks like many changes but most of them are same tabular alignments. We replaced the binary search with a slightly modified one (I restate the above):

arr = [ 1 ,2 ,2 ,2 ,2 ,2 ,2 ,2, 3 ,3 ,4 ]
key = 2
task: binary search to find the first position of 2
problem: a classical binary search would give back the 2 in the middle. But we are searching the one at the beginning
solution: keep doing the binary search until we hit the first position of 2 and not stopping on the first hit and going back iteratively.

With these changes we are achieving a similar performance for index creating like the c++ version.