@AlyHdr 2022-01-12
> Cool! If I understand correctly, this doesn't change anything to the HDT format itself, right?

Yes it's just a performance improvement...